requirements.txt
pydantic>=2.0
PyYAML
